Rename to internal-export to match dom-export
