+++ packages/react-router/lib/types/internal-export/route-data.ts
