+++ packages/react-router/lib/types/internal-export/route-module-annotations.ts
